samples/solace-samples-python/requirements.txt
autopep8==1.6.0
certifi==2020.12.5
chardet==4.0.0
idna==2.10
pycodestyle==2.8.0
requests==2.25.1
solace-pubsubplus==1.6.0
toml==0.10.2
urllib3==1.26.4
grpcio==1.60.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-api==1.22.0
